% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws_cli_install.R
\name{aws_cli_install}
\alias{aws_cli_install}
\title{Install AWS CLI}
\usage{
aws_cli_install()
}
\description{
This function download and install the AWS CLI for windows, linux or
mac. For Linux and Mac is necessary Python 2.6.5 or greater.
You can also install the CLI manually,
see \url{https://aws.amazon.com/cli/}
}
