% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configure.R
\name{aws_configs}
\alias{aws_configs}
\alias{aws_configure}
\alias{aws_configure_list}
\title{User configuration}
\usage{
aws_configure(profile = "default", access_key = NULL,
  secret_key = NULL, region = NULL)

aws_configure_list(profile = "default")
}
\arguments{
\item{profile}{The profile-name. If profile was not supplied
\code{aws_profile()} will create a default profile.}

\item{access_key}{The access key create by AWS.}

\item{secret_key}{The secret key create by AWS.}

\item{region}{The default region}
}
\description{
Functions to configure and visualize the user credentials
}
