% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws_version.R
\name{aws_version}
\alias{aws_version}
\title{AWS CLI version}
\usage{
aws_version()
}
\description{
This function provides the version of AWS CLI installed. A message
to use \code{aws_cli_install()} will be exhibit if the AWS CLI
was not found.
}
