% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_profile_settings.R
\name{change_settings}
\alias{change_settings}
\alias{change_access_key}
\alias{change_secret_key}
\alias{change_region}
\title{Change profile settings}
\usage{
change_access_key(access_key, profile = "default")

change_secret_key(secret_key, profile = "default")

change_region(region, profile = "default")
}
\arguments{
\item{access_key}{The access key create by AWS.}

\item{profile}{The profile-name. If profile was not supplied
\code{create_profile()} will create a default profile.}

\item{secret_key}{The secret key create by AWS.}

\item{region}{The default region}
}
\description{
Modify access key, secret key, and region in an existent profile.
}
\examples{
\dontrun{
# create a default user
create_profile(access_key = "my_access_key_1",
               secret_key = "123456789",
               region = "us-east-1")

# verify if the user was created
profile_settings()

# change region
change_region("us-west-1")

# verify if the region really change
profile_settings()

# remove your credentials from this computer
delete_all_profiles()
}

}
