% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_profile.R
\name{list_profiles}
\alias{list_profiles}
\title{The profile names saved}
\usage{
list_profiles(aws_file = "credentials")
}
\arguments{
\item{aws_file}{A string: "credentials" or "config".}
}
\value{
A tibble with the profile names available.
}
\description{
The AWs CLI creates two profile files: credentials
and config. Use \code{list_profiles("config")} if you want to see
the profile names in config file.
}
\examples{
 \dontrun{
 # use this for profile names available
 list_profiles()

 # use this if want to see the profiles on config file
 list_profiles("config")
 }
}
