% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_profiles.R
\name{list_profiles}
\alias{list_profiles}
\title{The profile names saved}
\usage{
list_profiles()
}
\value{
A tibble with the profile names available.
}
\description{
The AWs CLI creates two profile files: credentials
and config. This function return the profiles availables in both
files.
}
\examples{
 \dontrun{
 # use this for profile names available
 list_profiles()
 }
}
